// frontend/src/app/core/api/generated/models/asset-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AssetResponse {
  criticality?: 'Low' | 'Medium' | 'High' | 'Critical';
  id?: string;
  lastInspectedAt?: string;
  location?: string;
  name?: string;
  status?: 'Operational' | 'Degraded' | 'Maintenance' | 'Out of Service';
  type?: string;
}